Moved modal functions to own file
