+++ src/containers/MyNdla/Arena/ArenaButtons.tsx
@@ -6,21 +6,12 @@
  *
  */
 
-import { Dispatch, memo, SetStateAction, useCallback } from 'react';
-import { useLocation, useNavigate, useParams } from 'react-router-dom';
+import { Dispatch, memo, SetStateAction } from 'react';
+import { useLocation } from 'react-router-dom';
 import styled from '@emotion/styled';
 import { Plus } from '@ndla/icons/action';
-import { ArenaFormValues } from './components/ArenaForm';
 import ArenaTextModal from './components/ArenaTextModal';
-import { useCreateArenaTopic, useReplyToTopic } from '../arenaMutations';
-import {
-  arenaCategoryQuery,
-  arenaTopicById,
-  useArenaCategory,
-} from '../arenaQueries';
-
-const toArenaTopic = (topicId: number | undefined) =>
-  `/minndla/arena/topic/${topicId}`;
+import useArenaModal from '../useArenaModalActions';
 
 const StyledListItem = styled.li`
   margin: 0;
@@ -32,57 +23,8 @@ interface ArenaButtonsProps {
 }
 
 const ArenaButtons = ({ setFocusId, topicId }: ArenaButtonsProps) => {
-  const { categoryId } = useParams();
   const location = useLocation();
-  const navigate = useNavigate();
-  const { replyToTopic } = useReplyToTopic({
-    refetchQueries: [
-      {
-        query: arenaTopicById,
-        variables: { topicId, page: 1 },
-      },
-    ],
-  });
-
-  const { arenaCategory } = useArenaCategory({
-    variables: { categoryId: Number(categoryId), page: 1 },
-    skip: !Number(categoryId),
-  });
-
-  const { createArenaTopic } = useCreateArenaTopic({
-    refetchQueries: [
-      {
-        query: arenaCategoryQuery,
-        variables: { categoryId: arenaCategory?.id, page: 1 },
-      },
-    ],
-  });
-
-  const createReply = useCallback(
-    async (data: Partial<ArenaFormValues>) => {
-      const newReply = await replyToTopic({
-        variables: { topicId: Number(topicId), content: data.content ?? '' },
-      });
-      setFocusId?.(newReply.data?.replyToTopic.id);
-    },
-    [replyToTopic, topicId, setFocusId],
-  );
-
-  const createTopic = useCallback(
-    async (data: Partial<ArenaFormValues>) => {
-      if (arenaCategory) {
-        const topic = await createArenaTopic({
-          variables: {
-            content: data.content ?? '',
-            title: data.title ?? '',
-            categoryId: arenaCategory?.id,
-          },
-        });
-        navigate(toArenaTopic(topic.data?.newArenaTopic?.id));
-      }
-    },
-    [arenaCategory, createArenaTopic, navigate],
-  );
+  const { createReply, createTopic } = useArenaModal({ setFocusId, topicId });
 
   const showNewPostButton = location.pathname.includes('category');
   const showNewReplyButton = location.pathname.includes('topic');
